[FIX] #15 CoreData에 이미지 저장 로직 분리
