--- Pinit/Pinit/Repository/Repository.swift
@@ -1,9 +1,0 @@
-//
-//  Repository.swift
-//  Pinit
-//
-//  Created by 안정흠 on 3/12/25.
-//
-
-import Foundation
-
